Refactor Field to fv3fit.emulation.zhao_carr_fields
